requirements.txt
mne
pandas
numpy
matplotlib
seaborn
pyyaml
